chore

+++ src/main/java/staysplit/hotel_reservation/photo/domain/dto/request/UploadPhotoRequest.java
@@ -3,7 +3,7 @@ package staysplit.hotel_reservation.photo.domain.dto.request;
 // TODO:  LONG -> INTEGER
 public record UploadPhotoRequest(
         String entityType, // hotel or room
-        Long entityId,
+        Integer entityId,
         String dirName
 ) {
 }
